requirements.txt
streamlit>=1.28.0
googlesearch-python>=1.2.3
pandas>=2.0.0
wordcloud>=1.9.0
matplotlib>=3.7.0
urllib3>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0